As per testing with Kevin, mark this test as FIXME
The test is 'open a file in a project works and renders, open another file in different project with errors, it should clear the scene'

Kevin says it doesn't work because we never actually show KCL errors upon merely opening a file. We show them when we try to execute the file, which might depend on scroll position, viewport, something else.

Unclear exactly why this PR makes the test fail, but it shouldn't be passing on main at all. Main is a false positive. So I am marking this FIXME.