Test updates: identify line of interest properly
Previously, the line of interest could be identified via the stdlib fn
that was called. However now the test should really check both the
stdlib fn AND its parameters, because yLine could be either an xRelative
or xAbsolute constraint depending on which parameter is used, length or
endAbsolute.

So, instead of making the test more complex, I just added a comment
to the KCL example code being tested, and the test searches for that
comment instead.