Fix playwright test names
The playwright e2e test suite generated a bunch of playwright tests by
taking fragments of KCL code, each of which called a different sketch
function (e.g. yLineTo, yLine) and using that function as the test name.

Now that the functions have been changed to use kw args, there were two
tests with the same name (yLineTo which became yLine, and the old yLine).

This caused playwright to error out because it can't handle two tests
with the exact same name.

Fix was to change the name so it splits on = not on (. This should
generate tests with names like `yLine(endAbsolute` and `yLine(length`.