Fix removeSingleConstraint for angledLine calls
Problem was, previously the relevant code was looking for an object
property in the 0th arg, called 'angle' or 'length' or something. But
now the angledLine function is not using an object with a property.
Instead it's using multiple labeled arguments. So the code has to know
how to find and replace those arguments.

Unrelatedly, the tests had to be updated because now the function name
doesn't uniquely pick out a test case. We did this in previous commits
by using regexes. But we realized it's simpler to just add a comment to
search for.