Fix findUsesOfTagInPipe which was broken
It was looking for references to tags by finding kwarg labeled `tag=`.
But that's used for DECLARING tags, not for REFERENCING them.

Instead, it shouldn't really look at the kwarg labels at all. It should
just find tag declarations and references by their type.

This has a problem, in that if you assign a tag to an identifier, it
won't find them.

So... I guess... don't do that??? good enough for now.