Don't access `id` when editing a shell without a `VariableDeclaration`
The type on `getNodeFromPath` is doing us a disservice here, lying to us
when we actually don't have a `VariableDeclaration` in cases like the
one in `rust/kcl-lib/tests/fillet-and-shell/input.kcl`, which was
causing an error when attempting to edit via point-and-click. Now it
bails out of that, and falls back to giving that node a variable
declaration.