KCL: Getter for axes of planes
It's useful for geometry calculations in KCL to get the underlying axes and origin
of planes. This helps write transformations between planes for rotating geometry.

My goal was just to enable a simple getter for planes, like `myPlane.xAxis` and yAxis
and origins. That's nearly what happened, except I discovered that there's two ways
to represent a plane: either `KclValue::Plane` or `KclValue::Object` with the right fields.

No matter which format your plane is represented as, it should behave consistently
when you get its properties. Those properties should be returned as `[number; 3]`
because that is how KCL represents points.

Unfortunately we actually require planes-as-objects to be defined with axes like
`myPlane = { xAxis = { x = 1, y = 0, z = 0 }, ...}`, but that's a mistake in my opinion.
So if you do use that representation of a plane, it should still return a [number; 3].
This required some futzing around so that we let you access object fields .x and .y as [0]
and [1], which is weird, but whatever, I think it's good.

This PR is tested via planestuff.kcl which has a Rust unit test.

Part of the hole efforts, see #7543