KCL: Track Z axis of planes
KCL should track which way a plane is facing. If it only tracks a plane's
X and Y axes, it could identify two separate planes (overlapping
perfectly in space, but one plane's "up" is the other's down).

This is helpful for knowing which way to extrude on a given plane, or
which way to cut "into" a given solid.