Rework top bar to not include toolbar (#7681)
* Rework top bar to not include toolbar

Closes #7679 by creating a definite top bar, and moving the toolbar
below it. Still side-steps the E2E test issue by allowing the modeling
scene extend under this top bar. I will finally address this in the next
step, which is bringing back a proper sidebar that doesn't overlay the
modeling scene.

I have a fast-follow PR coming that adds visual undo and redo buttons to
this top bar, but I wanted to keep them separate.

* Fix conflict

* Rework top bar to not include toolbar

Closes #7679 by creating a definite top bar, and moving the toolbar
below it. Still side-steps the E2E test issue by allowing the modeling
scene extend under this top bar. I will finally address this in the next
step, which is bringing back a proper sidebar that doesn't overlay the
modeling scene.

I have a fast-follow PR coming that adds visual undo and redo buttons to
this top bar, but I wanted to keep them separate.

* Tiny CSS fix

* Update snapshots

* Remove circular dependency, avoid adding 2 new ones

* Fix up tests

* Patch up the last few tests

* Delete bad committed snapshot test result

* Update snapshots

* Remove AllMachines type

Sorry this doesn't have anything to do with the top bar but tsc is
blocking me because of this dumb utility type.

* Fix rectangle screenshot that broke because toolbar tooltips moved

Oy vey.

* Update snapshots

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>