KCL: More accurate benchmark for sketch mode perf (#7820)
The latency being discussed in https://github.com/KittyCAD/modeling-app/pull/7812 is actually latency of running a mock execution _with usePreviousMemory_ set to true. This means that the KCL prelude is only evaluated once. It's basically running with a "warm cache". So let's correct the benchmark, so it more accurately measures the interactive latency when users are dragging points around.