KCL: Preallocate space for all the paths
If you sketch a 100-sided shape, Rust will append 100 path segments to the underlying sketch.
Unfortunately because the vector which stores the paths starts at capacity 0, this could
cause quite a lot of allocations.

As an optimization, we will instead preallocate space for N paths, where N is the number
of steps in the |> pipeline. This will hopefully reduce allocations.