KCL: Rm clone in parser for 3% speedup (#7857)
Previously, building a member expression required cloning the Expr that
became the member expression's object. Now I changed things a bit, so
that the object expr doesn't need to be cloned.

This is used every time an Expr is parsed, so it's somewhat hot code,
and fixing the clone gave a consistent 2-4% speedup across tests on my
Macbook Pro.