Sketch on normal of face (#7886)
This work updates `startSketchOn` to enable sketching on a plane normal to the selected face. You can now specify an axis of the face that will become the x axis of the new plane, with the normal of the selected face becoming the y axis of the new plane. This extension creates the plane and begins sketching as usual.