KCL: Revolve should create new solids
When you extrude a sketch-on-face, you could either modify the original
solid, or create a new one. These are known as extrude methods (Merge
and New respectively).

Revolve, on the engine side, always does the equivalent of New for
revolve, but the app wrongly thought it was doing a Merge.

This meant KCL got confused about whether a revolved SoF should be treated
as its own object or part of something else's.