KCL: Revolve should create new solids (#7937)
When you extrude a sketch-on-face, you could either modify the original solid, or create a new one. These are known as extrude methods (Merge and New respectively).

Revolve, on the engine side, always does the equivalent of New for revolve, but the app wrongly thought it was doing a Merge.

This meant KCL got confused about whether a revolved SoF should be treated as its own object or part of something else's. Here's an easy way to visualize the problem:

Before:
<img width="1840" height="1037" alt="Screenshot 2025-07-30 at 2 07 42 PM" src="https://github.com/user-attachments/assets/f82f5107-1017-4b12-91f6-baf6ea3a6797" />

After:
<img width="1840" height="1151" alt="Screenshot 2025-07-30 at 2 07 10 PM" src="https://github.com/user-attachments/assets/1be72b33-a76b-47b0-9357-8583f86204fc" />

Fixes https://github.com/KittyCAD/modeling-app/issues/6925