KCL: Track profile open/close, better error on double-close (#7933)
Previously, trying to close an already-closed KCL sketch would give the error

```
engine: The provided object is of type Object, which is not a valid input for use in this operation.
```

That sucks! Terrible! Now it gives

```
This sketch is already closed. Remove this unnecessary `close()` call
```

Fixes https://github.com/KittyCAD/modeling-app/issues/7931