Docs: use 3D models for each KCL example (#8005)
The 3D models from https://github.com/KittyCAD/modeling-app/pull/7995 should be used in each KCL stdlib function's documentation page. So, I've updated the Markdown template to display them.

Once this is merged, it'll kick off a PR to documentation, which will kick off a PR to the website, which will actually display the 3D docs.