All this time our editor has used `monospace`....wow
Better late then never I guess! We use Source Code Pro elsewhere, so all
this PR does is apply it to the CodeMirror instance. Source Code Pro is
much more suitable for small font sizes, so I actually felt the need to
scale down the base font for the editor after changing it. It was a
little debatable but I went with `text-xs` for higher density. This has
made me realize that we really should make settings for the following
when we have time:
- [ ] setting for app font-size
- [ ] setting for editor font-size
- [ ] setting for editor font-family (maybe could take a URL?)