KCL docs: way to force fallback to 2D images
Currently the [KCL helix examples](https://zoo.dev/docs/kcl-std/functions/std-helix) do not properly export to GLTF.

We'll obviously need to fix that, but in the meantime, we will need a way to mark those tests as "use the 2D png for this example, not the 3D model".

If you add `no3d` to the start of a KCL example, its 3D export won't be
used and its 2D image will be preferred in the docs.