All this time our editor has used `monospace`. Use a real code font (#8076)
* All this time our editor has used `monospace`....wow

Better late then never I guess! We use Source Code Pro elsewhere, so all
this PR does is apply it to the CodeMirror instance. Source Code Pro is
much more suitable for small font sizes, so I actually felt the need to
scale down the base font for the editor after changing it. It was a
little debatable but I went with `text-xs` for higher density. This has
made me realize that we really should make settings for the following
when we have time:
- [ ] setting for app font-size
- [ ] setting for editor font-size
- [ ] setting for editor font-family (maybe could take a URL?)

* Update snapshots

* Actually add Source Code Variable font to the repo

* Move "Inter" font into a public/fonts/ dir

* Fix up CSP merge

* Update snapshots

* Update snapshots

* Fix loft E2E test that broke from text size decrease

Oy vey

* Fix E2E test that broke with font change

Unsure why

* Fix E2E test that broke because it clicked on some random code location

And deleting in another spot after the font size changed happened to not
cause the error this test expected to occur

* Update snapshots

* Update snapshots

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>