Add a code pane menu item to copy current file's code
A customer request to make it easier to copy a file's contents. Since we
have the menu here it is a very simple addition that shouldn't impact UX
for users who don't need it very often.

In the future we should add this into the potential shortcut actions
that just aren't assigned a keyboard shortcut by default, so power users
can wire it up to a hotkey sequence.