KCL: new 'hole' module for mech-eng  (#8530)
Add a new `hole` module with countersink, counterbore etc holes. This has been ready for two months or so, but I got distracted by constraint solvers. Clearly the ML team has found workarounds for converting holes in other CAD programs, but we should just add this as an experimental library and let people test it out in prod.