KCL: Disable pattern_into_union test (#8571)
Engine has a bug around this, https://github.com/KittyCAD/engine/issues/3865. We are disabling this test until engine fixes the issue, rather than reverting the engine PR which caused this to fail. Why? Because this test case was already triggering a bug in the engine, it just did so in a hidden and tricky way (copying the brep unnecessarily). We'd rather have this case fail with a loud error than quietly do inefficient brep copies.