KCL: Move flat-bottom hole check (#8569)
If the hole has a bottom drill angle of 180, it would cause a divide by zero error in our hole-calculation math. So I check for this and change it to 179.99

This check was only being done when the user explicitly selected a flat bottom. But if they put a bottom with an angle of 180, it wasn't checked. This PR runs the check no matter whether the user selected a `flat` bottom of a bottom with an angle of `180`.