Watch for a query parameter for the prompt within the text-to-cad UI (#8676)
* Watch for a query parameter for the prompt within the text-to-cad UI

All of the "try in browser" links in the website for TTC have been
broken since we migrated from the command-based implementation to the
pane-based implementation. This provides a way for URLs to include a
prompt that will be picked up by the pane UI and then deleted from the
search parameters. A separate PR is being made to allow URLs to set the
layout, so that we can guarantee that the TTC pane is open in a new
iteration of these "try in browser" links. But it's worth noting that
this implementation does not rely on the pane being open initially: the
param will stick around until the pane is opened.

* Hoist the state up and out of the textarea component

* Fix cursor-at-end problem

* fix: tsc errors