Use an auto-awaiting assertion on E2E test to deflake it (#8766)
Our test seems to have run the risk of presenting double toasts to the
user so we patched it using the Playwright `count()` assertion. However,
this assertion is instantaneous, so the export functionality which can
take a variable and unknown amount of time would often not be completed
when it was done. This replaces `count` with the normal `.toBeVisible`,
while keeping the previous test's lack of care for > 1 toasts by using
the `.first()` helper function on the locator.