fix: file extensions for imported files should work with and without capital letters (#8788)
* chore: writing some unit tests

* chore: updating the locations that implement the filter logic

* fix: saving off progress

* fix: fmt

* lowercase extension before detecting file format

For some filesystems, the casing of filenames can be all over the place.
For instance, FAT32 and APFS/HFS treat FOO.TXT as the same as foo.txt,
while other filesystems (extfs, btrfs, zfs, nfs, etc etc) treat them
as different files.

As a result, very few things care about the case of a file extension.
A .STEP file is the same as a .step file, and the filepath can get
mangled by round tripping through an uncooperative filesystem or
filesystem driver.

This will lowercase the extension before checking for known filetypes,
which is purely used for detection - not display or even loading from
the filesystem, in case we're on a filesystem where we are case
sensitive.

* chore: live computing the import list

* fix: fmt

* add test coverage for extensions

* lowercase path extension in the parser

* another extension via a cloned copy

* fix: cleaning up pr.

* fix: removing console log

* fix: this should fix it? I did the extension without period ope

* fix: does this fix it? I think insert should see kcl files.

* fix: YO

---------

Co-authored-by: Paul Tagliamonte <paultag@gmail.com>