Add PRINCIPLES.md
Add a doc where we can write down principles our team (and community)
agree on. This is in the spirit of trying to align on how we write code,
mock up workflow, define requirements, and work together as a team. I
imagine we could have this doc pulled up in meetings, referenced via URL
to a line number in a comment or review, or otherwise pointed to while
we build stuff together.

I proposed a few that I have on my mind right now, but give me feedback
at any level!