Part of https://github.com/KittyCAD/modeling-app/issues/8616
KCL's runtime needs to find the AST node for a given source range. This is called "pathToNode", and it's used by the frontend to figure out which solids or sketches correspond to a user's click, or a user's given variable.

Previously, in bug 8616, we describe a case where the pathToNode lookup fails, because the user used KCL's new `hole` module. This is written in KCL. In this case the pathToNode was trying to look up module 15, i.e. the stdlib `hole` module. But this failed because the holeToNode lookup only had the currently-open KCL file, i.e. module 0.

This PR fixes that problem. Now, looking up pathToNode will take _all_ KCL modules, not just the module for the current program. To do this, we introduce a new `ProgramLookup` struct which maps module IDs to the module contents. pathToNode lookups now use that type instead of just a single program.