Make common avatar style with rounded corners
Some user feedback from @lee-at-zoocorp that the avatar in the top bar
feels a bit bad, either because of the rounding, the border, or both. He
made the avatars in the TTC copilot chat have only slightly rounded
corners and no border. This makes a common tailwind class that both pull
from, and brings back a slight semitransparent border in case an avatar
image is close to the background color for legibility.