Add PRINCIPLES.md and GLOSSARY.md (#8884)
* Add PRINCIPLES.md

Add a doc where we can write down principles our team (and community)
agree on. This is in the spirit of trying to align on how we write code,
mock up workflow, define requirements, and work together as a team. I
imagine we could have this doc pulled up in meetings, referenced via URL
to a line number in a comment or review, or otherwise pointed to while
we build stuff together.

I proposed a few that I have on my mind right now, but give me feedback
at any level!

* Add GLOSSARY.md

* Remove a couple proposed principles

They're new proposals, and this PR should really be about documenting
what we have today, and just making a slot for these to live in.

* Add toolbar and orientation control sections

* Add images for the layouts at least?

I might regret adding images, because they'll fall out of date, meh

* @jamwaffles and @jacebrowning feedback

* Doh, misused "Workspace" thanks @graphite-bot

* Update workspace image and alt text