Create CodeMirror extension to register AST as StateField
Towards #8885, stacked on #9236. Adding a for-now unused CodeMirror
extension that registers the AST as a StateField on the EditorView.
Right now we trash the EditorView constantly so that value will not be
trustworthy, but it will be once we remove all our React scaffolding
around CodeMirror, which is a work-in-progress in #9190. Then this will
be available to become the canonical derived AST state, automatically
updated whenever the code is updated by the user.