Organize class properties on KclManager (#9236)
* Organize class properties on KclManager

Baby step towards #8885 broken out from a larger effort in #9190. Just
meant to start categorizing and commenting on the many class properties
in this class, as we step towards consolidating them.

* Update snapshots

* Update snapshots

* Revert AST setter

This contained material changes that break the app currently, relying on
other future changes.

* Update snapshots

* Update snapshots

* Update snapshots

* Revert changes to `code` setter

These rely on other changes to come to work properly.

* Update snapshots

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>