Create CodeMirror extension to register AST as StateField  (#9237)
* Organize class properties on KclManager

Baby step towards #8885 broken out from a larger effort in #9190. Just
meant to start categorizing and commenting on the many class properties
in this class, as we step towards consolidating them.

* Update snapshots

* Update snapshots

* Revert AST setter

This contained material changes that break the app currently, relying on
other future changes.

* Update snapshots

* Update snapshots

* Update snapshots

* Revert changes to `code` setter

These rely on other changes to come to work properly.

* Update snapshots

* Add CodeMirror extension to derive AST from code

* Create CodeMirror extension to register AST as StateField

Towards #8885, stacked on #9236. Adding a for-now unused CodeMirror
extension that registers the AST as a StateField on the EditorView.
Right now we trash the EditorView constantly so that value will not be
trustworthy, but it will be once we remove all our React scaffolding
around CodeMirror, which is a work-in-progress in #9190. Then this will
be available to become the canonical derived AST state, automatically
updated whenever the code is updated by the user.

* Forgot to revert where the extension gets registered to current setup

* Fix tsc

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>