Remove artifact decoration plugin (#9374)
Fixes #9366. This Decoration plugin highlights the code with
corresponding `artifactGraph` decorations whenever the debug panel is
active, and with certain KCL code it can get itself into an infinite
loop. This manifests as an internal back-and-forth within CodeMirror's
`sync()` methods, which call `super.sync()` and `child.sync()` endlessly
without settling. I don't understand why this plugin creates that
behavior, but it's not worth debugging fully until we want to turn it
back on.

I think this means that I should not advocate for using Decorator
plugins for any sort of state lookups or connecting tissue, which makes
sense in retrospect. Developers should use the StateFields and
calculated Facets to get one state representation they have to one they
want.