KCL: Track profile closed more carefully (#9381)
Previously the 'is_closed' field of a profile was
either true or false. I've changed it to an enum
with 4 states.

 - no: it's not closed, it's open.
 - yes, explicitly: user called close(), or made
   a closed shape e.g. circle, polygon
 - yes, implicitly: user made a line(endAbsolute)
   that went back to the sketch start
 - maybe: it's possible. The user added a relative
   line, and that line might have looped back on
   itself, KCL couldn't tell.

This is important to track because KCL doesn't
currently validate whether a profile is closed
before doing a solid extrude. We rely on the engine
to detect this.

However, the frontend team wants KCL to detect this,
so they can show the user an error message when they
select an open profile and a solid extrude. This
early validation and early error message will be
valuable.

This PR doesn't actually do that work, but it sets
the necessary refactor to make that work easy in a
followup PR.

Part of https://github.com/KittyCAD/modeling-app/issues/9367