Close a test sketch
This test previously worked because it's running
without an engine, but it wouldn't have worked in
production (because the sketch is open, and you can't
solid extrude an open sketch).

I'm not sure exactly what to do about this test, but
I think closing it will be fine so the test works.