Update PRINCIPLES.md regarding WASM method use
Principle 2 used to advise us to "reduce usage of initPromise" but as of #9308 that's no longer an issue, since it isn't exported ✨.

Instead, I'd like to use the slot to remind us to never repeat the pattern of importing methods from the WASM wrapper module directly. We create an instance of that module at runtime, and manage its lifecycle, potentially reloading it if it breaks. But it is confusing, because it's "just another JS module" that can be imported from while developing. Before #9415 we had helper methods that would import the methods directly, and take an instance of the module only optionally if at all. This could lead to unknown behavior if WASM needs reloaded or more.